make the actual journal be a subdirectory
A previous implementation had the journal be at the top-level and the
scripts be in a .journal directory, which required .bashrc and other
initialization scripts to be symlinked.

If we invert it so that the journal is a subdirectory we can keep the
initialization scripts at the top level and not have to symlink
anything.