bash_profile: automatically start vim when opening the journal
Not sure yet if this is what I want. However, it makes it easier to
open and close files, which will immediately close the journal and
commit the changes.

This will hide access to the shell a little bit, but it can still be
accessed via :sh inside vim.

To create a new file in vim using directory mode one needs to press %